Test if port is free, before using it for unit testing SFTP
* Also clean up code and reuse setting up SFTP server
