+++ filesystem/src/test/java/org/frankframework/filesystem/SftpFileSystemActorTest.java
@@ -1,18 +1,6 @@
 package org.frankframework.filesystem;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.Collections;
-import java.util.random.RandomGenerator;
-
-import org.apache.sshd.common.file.FileSystemFactory;
-import org.apache.sshd.common.file.virtualfs.VirtualFileSystemFactory;
 import org.apache.sshd.server.SshServer;
-import org.apache.sshd.server.auth.hostbased.StaticHostBasedAuthenticator;
-import org.apache.sshd.server.keyprovider.SimpleGeneratorHostKeyProvider;
-import org.apache.sshd.sftp.server.SftpSubsystemFactory;
 import org.frankframework.filesystem.ftp.SftpFileRef;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,9 +16,8 @@ class SftpFileSystemActorTest extends FileSystemActorTest<SftpFileRef, SftpFileS
 	private final String username = "frankframework";
 	private final String password = "pass_123";
 	private final String host = "localhost";
-	private int port = 22;
+	private static int port = 22;
 	private String remoteDirectory = "/home/frankframework/sftp";
-
 	private SshServer sshd;
 
 	@Override
@@ -39,41 +26,13 @@ class SftpFileSystemActorTest extends FileSystemActorTest<SftpFileRef, SftpFileS
 		if("localhost".equals(host)) {
 			remoteDirectory = "/"; // See getTestDirectoryFS(), '/' is the SFTP HOME directory.
 
-			sshd = createSshServer(username, password);
+			sshd = SftpFileSystemTestHelper.createSshServer(username, password, port);
 			log.info("Starting SSH daemon at port {}", sshd.getPort());
 			sshd.start();
 			port = sshd.getPort();
 		}
 
 		super.setUp();
-	}
-
-	static SshServer createSshServer(String username, String password) throws IOException {
-		SshServer sshd = SshServer.setUpDefaultServer();
-		sshd.setHost("localhost");
-		sshd.setPort(RandomGenerator.getDefault().nextInt(1024, 65535));
-		sshd.setPasswordAuthenticator((uname, psswrd, session) -> username.equals(uname) && password.equals(psswrd));
-		sshd.setHostBasedAuthenticator(new StaticHostBasedAuthenticator(true));
-
-		SftpSubsystemFactory sftpFactory = new SftpSubsystemFactory();
-		sshd.setSubsystemFactories(Collections.singletonList(sftpFactory));
-		sshd.setKeyPairProvider(new SimpleGeneratorHostKeyProvider());
-
-		sshd.setFileSystemFactory(getTestDirectoryFS());
-		return sshd;
-	}
-
-	/**
-	 * Creates the folder '../target/sftpTestFS' in which the tests will be executed.
-	 * This 'virtual FS' will pretend that the mentioned folder is the SFTP HOME directory.
-	 */
-	private static FileSystemFactory getTestDirectoryFS() throws IOException {
-		File targetFolder = new File(".", "target");
-		File sftpTestFS = new File(targetFolder.getCanonicalPath(), "sftpTestFS");
-		sftpTestFS.mkdir();
-		assertTrue(sftpTestFS.exists());
-
-		return new VirtualFileSystemFactory(sftpTestFS.toPath());
 	}
 
 	@Override
